requirements.txt
bs4
requests
discord.py